# mama-app-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mama na Mwana API Documentation",
        "description": "API documentation for Mama na Mwana application - A comprehensive maternal health tracking system",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/reminders/{user_id}": {
            "get": {
                "tags": [
                    "Reminders"
                ],
                "summary": "Get all reminders for a specific user",
                "description": "Returns a list of reminders for the specified user",
                "operationId": "getReminders",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to fetch reminders for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Reminder"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No reminders found for this user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No reminders found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reminders": {
            "post": {
                "tags": [
                    "Reminders"
                ],
                "summary": "Create a new reminder",
                "description": "Stores a new reminder for a user",
                "operationId": "storeReminder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "type",
                                    "reminder_time"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "doctor's appointment",
                                            "medicine",
                                            "medical tests"
                                        ],
                                        "example": "medicine"
                                    },
                                    "appointment": {
                                        "type": "string",
                                        "example": "visit therapist"
                                    },
                                    "reminder_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-03-21 14:30:00"
                                    },
                                    "dose_unit": {
                                        "type": "string",
                                        "enum": [
                                            "tablets",
                                            "drops",
                                            "capsule"
                                        ],
                                        "example": "tablets"
                                    },
                                    "medicine_details": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Vitamin D"
                                            },
                                            "dose": {
                                                "type": "string",
                                                "example": "2"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reminder created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reminder created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reminder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reminders/{id}": {
            "put": {
                "tags": [
                    "Reminders"
                ],
                "summary": "Update an existing reminder",
                "description": "Updates a reminder's information",
                "operationId": "updateReminder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the reminder to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reminder"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reminder updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reminder updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reminder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reminder not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reminder not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reminders"
                ],
                "summary": "Delete a reminder",
                "description": "Deletes a reminder by its ID",
                "operationId": "deleteReminder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the reminder to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reminder deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reminder deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reminder not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reminder not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Reminder": {
                "required": [
                    "user_id",
                    "type",
                    "reminder_time"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "doctor's appointment",
                            "medicine",
                            "medical tests"
                        ],
                        "example": "medicine"
                    },
                    "appointment": {
                        "type": "string",
                        "example": "visit therapist",
                        "nullable": true
                    },
                    "reminder_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-21 14:30:00"
                    },
                    "dose_unit": {
                        "type": "string",
                        "enum": [
                            "tablets",
                            "drops",
                            "capsule"
                        ],
                        "example": "tablets",
                        "nullable": true
                    },
                    "medicine_details": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "Vitamin D"
                            },
                            "dose": {
                                "type": "string",
                                "example": "2"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-21T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-21T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email/phone and password to get the authentication token",
                "name": "Token Based Authentication",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}